+++ mapcache/src/main/java/mil/nga/mapcache/GeoPackageManagerFragment.java
@@ -22,10 +22,6 @@ import android.os.Environment;
 import android.os.PowerManager;
 import android.provider.DocumentsContract.Document;
 import android.provider.Settings;
-import androidx.core.app.ActivityCompat;
-import androidx.core.content.ContextCompat;
-import androidx.core.content.FileProvider;
-import androidx.appcompat.app.AlertDialog;
 import android.text.InputFilter;
 import android.text.method.LinkMovementMethod;
 import android.view.LayoutInflater;
@@ -45,6 +41,11 @@ import android.widget.ListView;
 import android.widget.RadioButton;
 import android.widget.Spinner;
 import android.widget.TextView;
+
+import androidx.appcompat.app.AlertDialog;
+import androidx.core.app.ActivityCompat;
+import androidx.core.content.ContextCompat;
+import androidx.core.content.FileProvider;
 
 import com.ipaulpro.afilechooser.utils.FileUtils;
 
@@ -1365,6 +1366,8 @@ public class GeoPackageManagerFragment extends Fragment implements
         switch (table.getType()) {
 
             case FEATURE:
+                adapter.add(getString(R.string.geopackage_table_rename_label));
+                adapter.add(getString(R.string.geopackage_table_copy_label));
                 adapter.add(getString(R.string.geopackage_table_index_features_label));
                 adapter.add(getString(R.string.geopackage_table_create_feature_tiles_label));
                 adapter.add(getString(R.string.geopackage_table_add_feature_overlay_label));
@@ -1372,6 +1375,8 @@ public class GeoPackageManagerFragment extends Fragment implements
                 break;
 
             case TILE:
+                adapter.add(getString(R.string.geopackage_table_rename_label));
+                adapter.add(getString(R.string.geopackage_table_copy_label));
                 adapter.add(getString(R.string.geopackage_table_tiles_load_label));
                 adapter.add(getString(R.string.geopackage_table_link_label));
                 break;
@@ -1423,15 +1428,31 @@ public class GeoPackageManagerFragment extends Fragment implements
                         case 3:
                             switch (table.getType()) {
                                 case FEATURE:
+                                case TILE:
+                                    renameTableOption(table);
+                                    break;
+                            }
+
+                            break;
+                        case 4:
+                            switch (table.getType()) {
+                                case FEATURE:
+                                case TILE:
+                                    copyTableOption(table);
+                                    break;
+                            }
+                            break;
+                        case 5:
+                            switch (table.getType()) {
+                                case FEATURE:
                                     indexFeaturesOption(table);
                                     break;
                                 case TILE:
                                     loadTilesTableOption(table);
                                     break;
                             }
-
                             break;
-                        case 4:
+                        case 6:
                             switch (table.getType()) {
                                 case FEATURE:
                                     createFeatureTilesTableOption(table);
@@ -1441,14 +1462,14 @@ public class GeoPackageManagerFragment extends Fragment implements
                                     break;
                             }
                             break;
-                        case 5:
+                        case 7:
                             switch (table.getType()) {
                                 case FEATURE:
                                     addFeatureOverlayTableOption(table);
                                     break;
                             }
                             break;
-                        case 6:
+                        case 8:
                             switch (table.getType()) {
                                 case FEATURE:
                                     linkTableOption(table);
@@ -2092,6 +2113,113 @@ public class GeoPackageManagerFragment extends Fragment implements
                             }
                         }).create();
         deleteDialog.show();
+    }
+
+    /**
+     * Rename table option
+     *
+     * @param table
+     */
+    private void renameTableOption(final GeoPackageTable table) {
+
+        final EditText input = new EditText(getActivity());
+        input.setText(table.getName());
+
+        AlertDialog.Builder dialog = new AlertDialog.Builder(getActivity(), R.style.AppCompatAlertDialogStyle)
+                .setTitle(getString(R.string.geopackage_rename_label) + " " + table.getName())
+                .setView(input)
+                .setPositiveButton(getString(R.string.button_ok_label),
+                        new DialogInterface.OnClickListener() {
+                            public void onClick(DialogInterface dialog,
+                                                int whichButton) {
+                                String value = input.getText().toString();
+                                if (value != null && !value.isEmpty() && !value.equalsIgnoreCase(table.getName())) {
+                                    try {
+                                        active.removeTable(table);
+                                        GeoPackage geoPackage = manager.open(table
+                                                .getDatabase());
+                                        try {
+                                            geoPackage.renameTable(table.getName(), value);
+                                        } catch (Exception e) {
+                                            GeoPackageUtils.showMessage(getActivity(),
+                                                    "Rename " + table.getDatabase()
+                                                            + " " + table.getName()
+                                                            + " Table", e.getMessage());
+                                        } finally{
+                                            geoPackage.close();
+                                        }
+                                        update();
+                                    } catch (Exception e) {
+                                        GeoPackageUtils.showMessage(
+                                                getActivity(), "Rename "
+                                                        + table.getName(), e.getMessage());
+                                    }
+                                }
+                            }
+                        })
+                .setNegativeButton(getString(R.string.button_cancel_label),
+                        new DialogInterface.OnClickListener() {
+                            public void onClick(DialogInterface dialog,
+                                                int whichButton) {
+                                dialog.cancel();
+                            }
+                        });
+
+        dialog.show();
+
+    }
+
+    /**
+     * Copy table option
+     *
+     * @param table
+     */
+    private void copyTableOption(final GeoPackageTable table) {
+
+        final EditText input = new EditText(getActivity());
+        input.setText(table.getName() + "_copy");
+
+        AlertDialog.Builder dialog = new AlertDialog.Builder(getActivity(), R.style.AppCompatAlertDialogStyle)
+                .setTitle(getString(R.string.geopackage_copy_label) + " " + table.getName())
+                .setView(input)
+                .setPositiveButton(getString(R.string.button_ok_label),
+                        new DialogInterface.OnClickListener() {
+                            public void onClick(DialogInterface dialog,
+                                                int whichButton) {
+                                String value = input.getText().toString();
+                                if (value != null && !value.isEmpty() && !value.equalsIgnoreCase(table.getName())) {
+                                    try {
+                                        GeoPackage geoPackage = manager.open(table
+                                                .getDatabase());
+                                        try {
+                                            geoPackage.copyTable(table.getName(), value);
+                                        } catch (Exception e) {
+                                            GeoPackageUtils.showMessage(getActivity(),
+                                                    "Copy " + table.getDatabase()
+                                                            + " " + table.getName()
+                                                            + " Table", e.getMessage());
+                                        } finally{
+                                            geoPackage.close();
+                                        }
+                                        update();
+                                    } catch (Exception e) {
+                                        GeoPackageUtils.showMessage(
+                                                getActivity(), "Copy "
+                                                        + table.getName(), e.getMessage());
+                                    }
+                                }
+                            }
+                        })
+                .setNegativeButton(getString(R.string.button_cancel_label),
+                        new DialogInterface.OnClickListener() {
+                            public void onClick(DialogInterface dialog,
+                                                int whichButton) {
+                                dialog.cancel();
+                            }
+                        });
+
+        dialog.show();
+
     }
 
     /**